+++ src/app/admin/products/_components/ProductActions.tsx
@@ -5,7 +5,7 @@ import { useTransition } from "react";
 import {
   deleteProduct,
   toggleProductAvailability,
-} from "../../_actions/products";
+} from "../../../../utils/actions/products";
 import { useRouter } from "next/navigation";
 
 export function ActiveToggleDropdownItem({
