+++ src/app/admin/products/_components/ProductForm.tsx
@@ -6,7 +6,7 @@ import { Label } from "@/components/ui/label";
 import { Textarea } from "@/components/ui/textarea";
 import { formatCurrency } from "@/lib/formatters";
 import { useState } from "react";
-import { addProduct, updateProduct } from "../../_actions/products";
+import { addProduct, updateProduct } from "../../../../utils/actions/products";
 import { useFormState, useFormStatus } from "react-dom";
 import { Product } from "@prisma/client";
 import Image from "next/image";
