+++ src/hooks/checkout/useCheckoutForm.ts
@@ -7,7 +7,7 @@ import { toast } from "sonner";
 
 import { useCartStore } from "@/store/useCartStore";
 import { checkoutDetailsSchema } from "@/lib/validations";
-import { createOrder } from "@/app/admin/_actions/order.actions";
+import { createOrder } from "@/utils/actions/order.actions";
 import { calculateTotals, clearCart } from "@/utils/checkoutHelpers";
 
 export const useCheckoutForm = () => {
